+++ src/main/java/com/authlete/hms/fapi/FapiResourceRequestVerifier.java
@@ -17,9 +17,13 @@ package com.authlete.hms.fapi;
 
 
 import java.security.SignatureException;
+import java.util.ArrayList;
+import java.util.List;
+import com.authlete.hms.ComponentIdentifier;
 import com.authlete.hms.SignatureBase;
 import com.authlete.hms.SignatureEntry;
 import com.authlete.hms.SignatureMetadata;
+import com.authlete.hms.VerificationInfo;
 import com.authlete.hms.impl.JoseHttpVerifier;
 import com.nimbusds.jose.jwk.JWK;
 
@@ -122,7 +126,8 @@ public class FapiResourceRequestVerifier extends FapiResourceRequestBase<FapiRes
      *         A signature entry that contains the signature to be verified.
      *
      * @return
-     *         The result of signature verification.
+     *         Information about the verification operation, including the
+     *         computed signature base and the verification result.
      *
      * @throws IllegalStateException
      *         Mandatory input parameters, such as {@code method},
@@ -132,7 +137,8 @@ public class FapiResourceRequestVerifier extends FapiResourceRequestBase<FapiRes
      * @throws SignatureException
      *         Signature verification failed.
      */
-    public boolean verify(SignatureEntry signatureEntry) throws IllegalStateException, SignatureException
+    public VerificationInfo verify(SignatureEntry signatureEntry)
+            throws IllegalStateException, SignatureException
     {
         return verify(signatureEntry.getSignature(), signatureEntry.getMetadata());
     }
@@ -151,7 +157,8 @@ public class FapiResourceRequestVerifier extends FapiResourceRequestBase<FapiRes
      *         has been provided together with the signature should be used.
      *
      * @return
-     *         The result of signature verification.
+     *         Information about the verification operation, including the
+     *         computed signature base and the verification result.
      *
      * @throws IllegalStateException
      *         Mandatory input parameters, such as {@code method},
@@ -161,7 +168,8 @@ public class FapiResourceRequestVerifier extends FapiResourceRequestBase<FapiRes
      * @throws SignatureException
      *         Signature verification failed.
      */
-    public boolean verify(byte[] signature, SignatureMetadata metadata) throws IllegalStateException, SignatureException
+    public VerificationInfo verify(byte[] signature, SignatureMetadata metadata)
+            throws IllegalStateException, SignatureException
     {
         // Check if a verification key is set.
         checkVerificationKey();
@@ -169,8 +177,40 @@ public class FapiResourceRequestVerifier extends FapiResourceRequestBase<FapiRes
         // Create the signature base.
         SignatureBase base = createSignatureBase(metadata);
 
-        // Verify the signature.
-        return verify(signature, base);
+        // The verification result.
+        boolean verified = false;
+
+        // The reason of the verification result.
+        String reason = null;
+
+        try
+        {
+            // Check the metadata.
+            checkMetadata(base.getParamsLine().getMetadata());
+
+            // Verified the signature.
+            verified = verify(signature, base);
+
+            if (!verified)
+            {
+                reason = "The signature differs from the expected value.";
+            }
+        }
+        catch (Exception cause)
+        {
+            reason = cause.getMessage();
+        }
+
+        // Collect information about the verification operation.
+        VerificationInfo info = new VerificationInfo()
+                .setVerificationKey(getVerificationKey())
+                .setSignatureBase(base)
+                .setSignature(signature)
+                .setVerified(verified)
+                .setReason(reason)
+                ;
+
+        return info;
     }
 
 
@@ -199,5 +239,52 @@ public class FapiResourceRequestVerifier extends FapiResourceRequestBase<FapiRes
     {
         // Verify the signature using the specified verification key.
         return base.verify(new JoseHttpVerifier(getVerificationKey()), signature);
+    }
+
+
+    private void checkMetadata(SignatureMetadata metadata) throws SignatureException
+    {
+        checkComponentIdentifiers(metadata);
+        checkMetadataParameters(metadata);
+    }
+
+
+    private void checkComponentIdentifiers(SignatureMetadata metadata) throws SignatureException
+    {
+        List<ComponentIdentifier> requiredComponents = new ArrayList<>();
+
+        // "@method"
+        requiredComponents.add(COMP_ID_METHOD);
+
+        // "@target-uri"
+        requiredComponents.add(COMP_ID_TARGET_URI);
+
+        // "authorization"
+        requiredComponents.add(COMP_ID_AUTHORIZATION);
+
+        if (getDpop() != null)
+        {
+            // "dpop"
+            requiredComponents.add(COMP_ID_DPOP);
+        }
+
+        if (getContentDigest() != null)
+        {
+            // "content-digest"
+            requiredComponents.add(COMP_ID_CONTENT_DIGEST);
+        }
+
+        // Confirm that the metadata contains all the required components.
+        FapiResourceVerificationUtility.checkComponents(metadata, requiredComponents);
+    }
+
+
+    private void checkMetadataParameters(SignatureMetadata metadata) throws SignatureException
+    {
+        // created
+        FapiResourceVerificationUtility.checkParameterCreated(metadata);
+
+        // tag
+        FapiResourceVerificationUtility.checkParameterTag(metadata, TAG_VALUE_FAPI_2_REQUEST);
     }
 }