+++ src/test/java/com/authlete/hms/fapi/FapiResourceRequestTest.java
@@ -22,14 +22,16 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.net.URI;
 import java.security.SignatureException;
 import java.text.ParseException;
+import java.time.Duration;
 import java.time.Instant;
 import java.util.Arrays;
 import java.util.Base64;
 import org.junit.jupiter.api.Test;
 import com.authlete.hms.ComponentIdentifier;
-import com.authlete.hms.SignatureInfo;
 import com.authlete.hms.SignatureMetadata;
 import com.authlete.hms.SignatureMetadataParameters;
+import com.authlete.hms.SigningInfo;
+import com.authlete.hms.VerificationInfo;
 import com.nimbusds.jose.jwk.JWK;
 
 
@@ -85,13 +87,13 @@ public class FapiResourceRequestTest
         FapiResourceRequestSigner signer = createSigner(created, signingKey);
 
         // Sign
-        SignatureInfo info = signer.sign();
+        SigningInfo sinfo = signer.sign();
 
         // Check the signature metadata.
-        checkDefaultMetadata(created, info);
+        checkDefaultMetadata(created, sinfo);
 
         // Check the signature serialization.
-        checkSerialization(info);
+        checkSerialization(sinfo);
 
         // Create a verifier.
         FapiResourceRequestVerifier verifier = createVerifier(verificationKey);
@@ -100,21 +102,21 @@ public class FapiResourceRequestTest
         sleep(1000);
 
         // Verify with the default signature metadata. This verification should fail.
-        boolean verified = verifier.verify(info.getSignature(), null);
-        assertFalse(verified, "Signature verification unexpectedly passed.");
+        VerificationInfo vinfo = verifier.verify(sinfo.getSignature(), null);
+        assertFalse(vinfo.isVerified(), "Signature verification unexpectedly passed.");
 
         // Let the verifier use the 'created' value when it builds the default
         // signature metadata.
         verifier.setCreated(created);
 
         // Verify with the default signature metadata. This verification should pass.
-        verified = verifier.verify(info.getSignature(), null);
-        assertTrue(verified, "Signature verification unexpectedly failed.");
+        vinfo = verifier.verify(sinfo.getSignature(), null);
+        assertTrue(vinfo.isVerified(), "Signature verification unexpectedly failed.");
 
         // Verify with the same signature metadata as used for signing.
         // This verification should pass.
-        verified = verifier.verify(info.getSignature(), info.getMetadata());
-        assertTrue(verified, "Signature verification unexpectedly failed.");
+        vinfo = verifier.verify(sinfo.getSignature(), sinfo.getMetadata());
+        assertTrue(vinfo.isVerified(), "Signature verification unexpectedly failed.");
     }
 
 
@@ -143,7 +145,7 @@ public class FapiResourceRequestTest
     }
 
 
-    private static void checkDefaultMetadata(Instant created, SignatureInfo info)
+    private static void checkDefaultMetadata(Instant created, SigningInfo info)
     {
         // Expected signature metadata
         String expectedMetadata = String.format(
@@ -158,7 +160,7 @@ public class FapiResourceRequestTest
     }
 
 
-    private static void checkSerialization(SignatureInfo info)
+    private static void checkSerialization(SigningInfo info)
     {
         // Expected signature serialization
         String expectedSerializedSignature = String.format(":%s:",
@@ -194,26 +196,26 @@ public class FapiResourceRequestTest
                 );
 
         // Sign
-        SignatureInfo info = signer.sign(metadata);
+        SigningInfo sinfo = signer.sign(metadata);
 
         // Check the signature metadata.
-        checkCustomMetadata(created, info);
+        checkCustomMetadata(created, sinfo);
 
         // Create a verifier.
         FapiResourceRequestVerifier verifier = createVerifier(verificationKey);
 
         // Verify with the custom metadata. This verification should pass.
-        boolean verified = verifier.verify(info.getSignature(), metadata);
-        assertTrue(verified, "Signature verification unexpectedly failed.");
+        VerificationInfo vinfo = verifier.verify(sinfo.getSignature(), metadata);
+        assertTrue(vinfo.isVerified(), "Signature verification unexpectedly failed.");
 
         // Verify with the default metadata. This verification should fail.
         verifier.setCreated(created);
-        verified = verifier.verify(info.getSignature(), null);
-        assertFalse(verified, "Signature verification unexpectedly passed.");
+        vinfo = verifier.verify(sinfo.getSignature(), null);
+        assertFalse(vinfo.isVerified(), "Signature verification unexpectedly passed.");
     }
 
 
-    private static void checkCustomMetadata(Instant created, SignatureInfo info)
+    private static void checkCustomMetadata(Instant created, SigningInfo info)
     {
         // Expected signature metadata
         String expectedMetadata = String.format(
@@ -224,5 +226,120 @@ public class FapiResourceRequestTest
         String actualMetadata = info.getSerializedSignatureMetadata();
 
         assertEquals(expectedMetadata, actualMetadata);
+    }
+
+
+    @Test
+    public void test_missing_component() throws ParseException, IllegalStateException, SignatureException
+    {
+        JWK signingKey      = JWK.parse(SIGNING_KEY);
+        JWK verificationKey = signingKey.toPublicJWK();
+        Instant created     = Instant.now();
+
+        // Create a signer.
+        FapiResourceRequestSigner signer = createSigner(created, signingKey);
+
+        // Signature metadata that is missing mandatory components.
+        SignatureMetadata metadata = new SignatureMetadata(
+                // "@target-uri" is missing.
+                Arrays.asList(
+                        new ComponentIdentifier("@method"),
+                        new ComponentIdentifier("authorization"),
+                        new ComponentIdentifier("content-digest")
+                ),
+                new SignatureMetadataParameters()
+                    .setCreated(created)
+                    .setTag("fapi-2-request")
+        );
+
+        // Sign with the invalid metadata.
+        SigningInfo sinfo = signer.sign(metadata);
+
+        // Create a verifier.
+        FapiResourceRequestVerifier verifier = createVerifier(verificationKey);
+
+        // Verify with the same signature metadata as used for signing.
+        //
+        // This verification should fail because the required component
+        // "@target-uri" is missing.
+        VerificationInfo vinfo = verifier.verify(sinfo.getSignature(), sinfo.getMetadata());
+        assertFalse(vinfo.isVerified(), "Signature verification unexpectedly passed.");
+    }
+
+
+    @Test
+    public void test_expired() throws ParseException, IllegalStateException, SignatureException
+    {
+        JWK signingKey      = JWK.parse(SIGNING_KEY);
+        JWK verificationKey = signingKey.toPublicJWK();
+        Instant created     = Instant.now();
+
+        // Create a signer.
+        FapiResourceRequestSigner signer = createSigner(created, signingKey);
+
+        // Signature metadata with a 'created' parameter set too far in the past,
+        // causing the verifier to consider the HTTP message signature expired.
+        SignatureMetadata metadata = new SignatureMetadata(
+                Arrays.asList(
+                        new ComponentIdentifier("@method"),
+                        new ComponentIdentifier("@target-uri"),
+                        new ComponentIdentifier("authorization"),
+                        new ComponentIdentifier("content-digest")
+                ),
+                new SignatureMetadataParameters()
+                    .setCreated(created.minus(Duration.ofHours(1)))
+                    .setTag("fapi-2-request")
+        );
+
+        // Sign with the invalid metadata.
+        SigningInfo sinfo = signer.sign(metadata);
+
+        // Create a verifier.
+        FapiResourceRequestVerifier verifier = createVerifier(verificationKey);
+
+        // Verify with the same signature metadata as used for signing.
+        //
+        // This verification should fail because the 'created' parameter
+        // indicates that the HTTP message signature has expired.
+        VerificationInfo vinfo = verifier.verify(sinfo.getSignature(), sinfo.getMetadata());
+        assertFalse(vinfo.isVerified(), "Signature verification unexpectedly passed.");
+    }
+
+
+    @Test
+    public void test_bad_tag() throws ParseException, IllegalStateException, SignatureException
+    {
+        JWK signingKey      = JWK.parse(SIGNING_KEY);
+        JWK verificationKey = signingKey.toPublicJWK();
+        Instant created     = Instant.now();
+
+        // Create a signer.
+        FapiResourceRequestSigner signer = createSigner(created, signingKey);
+
+        // Signature metadata whose 'tag' parameter holds an unexpected value.
+        SignatureMetadata metadata = new SignatureMetadata(
+                Arrays.asList(
+                        new ComponentIdentifier("@method"),
+                        new ComponentIdentifier("@target-uri"),
+                        new ComponentIdentifier("authorization"),
+                        new ComponentIdentifier("content-digest")
+                ),
+                new SignatureMetadataParameters()
+                    .setCreated(created)
+                    .setTag("unknown")
+        );
+
+        // Sign with the invalid metadata.
+        SigningInfo sinfo = signer.sign(metadata);
+
+        // Create a verifier.
+        FapiResourceRequestVerifier verifier = createVerifier(verificationKey);
+
+        // Verify with the same signature metadata as used for signing.
+        //
+        // This verification should fail because the 'tag' parameter
+        // holds an unexpected value.
+        VerificationInfo vinfo = verifier.verify(sinfo.getSignature(), sinfo.getMetadata());
+        assertFalse(vinfo.isVerified(), "Signature verification unexpectedly passed.");
     }
 }